Feat: Multiplatform App View
